hard drop and optimised gameloop
prevented the game from drawing the entire board every game loop. As the game progresses and more shapes are placed on the board the game slowed down due to having to draw every shape, every frame, the board is now only drawn every time the game loop is divisible by a set number.